requirements.txt
# Requirements for Kabir Doha Chrome Extension
# Install with: pip install -r requirements.txt

Pillow>=11.0.0
